refactoring sharding package: - move everything related to Nodes Coordinator, to a separate package inside sharding
