+++ sharding/nodesCoordinator/indexHashedNodesCoordinator.go
@@ -1,4 +1,4 @@
-package sharding
+package nodesCoordinator
 
 import (
 	"bytes"
