+++ sharding/nodesCoordinator/shardingArgs.go
@@ -1,4 +1,4 @@
-package sharding
+package nodesCoordinator
 
 import (
 	"github.com/ElrondNetwork/elrond-go-core/data/endProcess"
